Let `flow` be opt-in to babel6.
